requirements.txt
pyrogram>=2.0.106
TgCrypto>=1.2.5
python-decouple>=3.8
motor>=3.3.2
pymongo>=4.6.1
aiohttp>=3.9.1
aiofiles>=23.2.0
anthropic>=0.8.1
openai>=1.6.1
cohere>=4.39
google-generativeai>=0.3.2
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3